fix(discord-plays-pokemon): dedupe the worker-boot spawn helper across tests
Why:
The jscpd ratchet failed CI: the new legacy-layout test's spawn-and-collect
block was an exact 118-token clone of the existing boot-readiness
compatibility test's block, both piping a worker source into `bun run -`
and collecting stdout/stderr/exitCode.

What:
Extract bootWorkerSource() into benchmark-worker-test-helpers.ts and use it
from both benchmark-harness.test.ts and benchmark-worker-legacy-layout.test.ts.

Verification:
- bunx turbo run typecheck lint test --filter=@discord-plays-pokemon/backend
  - all clean, 442 tests passed
- bun scripts/jscpd-check.ts - ratchet passed (583 clones, was 584)
