+++ packages/discord-plays-pokemon/packages/backend/src/goal/benchmark/benchmark-worker-test-helpers.ts
@@ -1,0 +1,20 @@
+/** Pipes workerSource into a fresh `bun run -` in cwd and collects its output. */
+export async function bootWorkerSource(
+  cwd: string,
+  workerSource: string,
+): Promise<{ output: string; exitCode: number }> {
+  const child = Bun.spawn(["bun", "run", "-"], {
+    cwd,
+    stdin: "pipe",
+    stdout: "pipe",
+    stderr: "pipe",
+  });
+  await child.stdin.write(workerSource);
+  await child.stdin.end();
+  const [stdout, stderr, exitCode] = await Promise.all([
+    new Response(child.stdout).text(),
+    new Response(child.stderr).text(),
+    child.exited,
+  ]);
+  return { output: stdout + stderr, exitCode };
+}
